RHEL and CentOS script fixes
This includes fixes for RHEL and CentOS installation scripts,
and README files. Includes a new way of checking for versions
that works for both types of /etc/redhat-release output

In addition, CentOS 7.5 has moved to the CentOS archive vault,
so we need to pull updates from a different place now.