Allow exit after build, take option to package
Modify the source build/install scripts to take an option
to exit after completing a build. This will leave code
behind in temp directories (why would you build and then
just throw away everything?).  This can be useful if we
want to test a build without trying to install it anywhere.

Set up the arguments to allow us to make packages instead
of directly installing the software.