Update tools to perform .deb and .rpm packaging
Add packaging to all of the component scripts, so that we can
recreate .rpm and .deb packages on RedHat-based and Debian-based
distros, respectively. This can be done with the '-p' option for
the source install scripts.

As part of this, we are changing the numbering scheme to split
up scripts for system tools (01) and libraries (03) so that it's
easier to add in new system tools without renumbering everything.

In addition, adding build scripts for the meta-packages that don't
build any software, but which are used for dependencies.