Fix updates for rocSPARSE and rocALUTION
We need to fix some hard-coded file paths in the rocSPARSE and
rocALUTION builds. However, the way this was done previously,
we might fail to do this change properly if we build a second or
more time. So before doing the change, checkout the original
version so we know we are working on the right thing.