Better handling of rocm-dkms deb from source
We previously missed the postinst file for rocm-dkms which set
the udev command allowing users access to the GPU if they were
in the video group. Make sure our .deb builds from source include
these files.

In addition, if we are "make installing" the meta-packages instead
of building the packages, set those udev rules as if the package
were "installed".