Add Python as a requirement for ROCm-SMI build
In case the user wants to install the rocm-smi package after
building it, it may rely on Python. New versions of rpm-build
will automatically figure this dependence out and add it to
the package. As such, install Python as a pre-req for rocm-smi.

(cherry picked from commit 144dfa0597f622e6401b86d54949d1ba2ed534d6)