Fix broken directory symlink handling
Doing X/{directory_symlink/../ in bash does not necessarily lead
you back to X. These scripts were broken because they were trying
to copy files from a parent directory, though the directory they
exist in is actually a symlink to somewhere else. As such, they
would fail to copy. This was not found because testing was done on
a system after doing a symlink-flattening copy with scp. Oops.

pushd and cd do handle such symlink things as we would like them
to, so instead of copying from {symlink}/../, we instead temporarily
change directories into {symlink}/../