Fix crash on attempting to reboot in a container
If you attempt to reboot while you're in a container, the command
will fail. Rather than just having our script blow up when they
try this, print out an easy to digest message explaining what
happened.

Running the shutdown command in a command-substitution chunk
prevents the "kill the script on any error" logic from kicking in
in this case.