Workaround for #4
In order to "recreate" what AMD ships in their binary packages,
for now we can force the HCC target cmake files to claim the
library IMPORTED_LOCATION stuff is all for 'release'. Even when
we are in a debug build. This is not a great fix, but it is a
reasonable workaround so that all sotware that is designed to
build against shipping ROCm will still work.