Fixes #6 and corrects a reboot mistake
The previous commit that tried to fix rebooting inside a container
caused rebooting to fail on RH-based distros (which do not use
/sbin/reboot) and also on real Ubuntu systems (because of a typo).
This fixes that problem.

It also solves #6 by checking to see if we have a chance of
correctly building the ROCK package against the currently installed
kernel. It does this by checking to see if the kernel development
headers for the running kernel are availabl on this system. If not,
we ask the user if we should just skip installing.