Fix how ATMI is build on ROCm 2.0
Previously, on ROCm 2.0, we were not building the device runtime
because a change to the ROCm device libs caused a required file to
no longer be installed. This adds a workaround for that.

Also adds a minor fix to how we will try to build LLVM for comgr
if ROCm LLVM wasn't built yet. Hadn't updated the script name for
for the tool to build LLVM.