Patch ROCr to prevent increase in library version
ROCr library version when built from GitHub is based on the
GitHub tags. These tags increase in version every time we do a
ROCm release, so that we can figure out which commit was used for
each ROCm release.

Unfortunately, that means that when we bumped to ROCm 2.0, we
suddenly started calling built-from-github ROCr shared objects
so.2 rather than so.1.

However, when we install from .deb or .rpm, these are still
numbered as .1. This could cause incompatibility between
software that has linked against the binary installed files but
wants to run on a system with the GitHub-sourced builds.

This patch hard-codes the build of ROCr to be a meaningful number
that is still in the "1.x" range, but is higher than what we
built from ROCm 1.9.x.